+++ package.json
@@ -13,10 +13,10 @@
     "test": "jest --collect-coverage",
     "test:dev": "jest --watch",
     "test:coverage": "jest --coverage",
-    "prisma:generate": "npx prisma generate --schema ./src/server/models/schema.prisma",
-    "prisma:migrate": "npx prisma migrate dev --schema ./src/server/models/schema.prisma",
-    "prisma:reset": "npx prisma migrate reset --schema ./src/server/models/schema.prisma --force",
-    "prisma:deploy": "npx prisma migrate deploy --schema ./src/server/models/schema.prisma",
+    "prisma:generate": "npx prisma generate --schema ./src/server/models/db/schema.prisma",
+    "prisma:migrate": "npx prisma migrate dev --schema ./src/server/models/db/schema.prisma",
+    "prisma:reset": "npx prisma migrate reset --schema ./src/server/models/db/schema.prisma --force",
+    "prisma:deploy": "npx prisma migrate deploy --schema ./src/server/models/db/schema.prisma",
     "prepare:husky": "husky install && npx husky add .husky/pre-commit \"npx lint-staged\" && npx husky add .husky/pre-push \"npm run type-check\"",
     "form-runner:install": "./src/form-runner/install-form-runner.sh",
     "form-runner:start": "cd ./src/form-runner/form-runner-app && yarn runner start"
@@ -39,7 +39,6 @@
     "node-fetch": "^2.6.1",
     "nunjucks": "^3.2.3",
     "pg": "^8.6.0",
-    "prisma": "^2.21.2",
     "sqlstring": "^2.3.2",
     "winston": "^3.3.3"
   },
@@ -82,6 +81,7 @@
     "nodemon-webpack-plugin": "^4.5.2",
     "postcss-loader": "^5.2.0",
     "prettier": "2.2.1",
+    "prisma": "^2.21.2",
     "sass": "^1.32.10",
     "sass-loader": "^11.0.1",
     "supertest": "^6.1.3",
