+++ src/server/models/helpers.ts
@@ -1,7 +1,7 @@
 import { upperFirst, isNumber } from "lodash";
 import { logger } from "server/services/logger";
-import { db } from "./database";
-import { CountriesWithData } from "./types";
+import { db } from "./db/database";
+import { CountriesWithData, CountryName } from "./types";
 
 const countriesWithData: CountriesWithData[] = [
   "Thailand",
@@ -10,8 +10,10 @@ const countriesWithData: CountriesWithData[] = [
   "Spain",
 ];
 
-export const countryHasLawyers = (countryName: string): boolean => {
-  return countriesWithData.includes(upperFirst(countryName));
+export const countryHasLawyers = (countryName: CountryName): boolean => {
+  return countriesWithData.includes(
+    upperFirst(countryName) as CountriesWithData
+  );
 };
 
 export const rawInsertGeoLocation = async (
