+++ src/server/models/types.ts
@@ -1,7 +1,10 @@
 import Prisma from "@prisma/client";
-import { countriesList } from "server/services/metadata"
+import { countriesList } from "server/services/metadata";
 
-export type CountriesWithData = Extract<CountryName, "Thailand" | "France" | "Italy" | "Spain">;
-export type CountryName = typeof countriesList[number]["text"];
+export type CountriesWithData = Extract<
+  CountryName,
+  "Thailand" | "France" | "Italy" | "Spain"
+>;
+export type CountryName = typeof countriesList[number]["value"];
 export type Point = number[];
 export type Lawyer = Prisma.Lawyer;
