+++ src/server/routes/development.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import { exec } from "child_process";
-import { prisma } from "server/models/prisma-client";
+import { prisma } from "server/models/db/prisma-client";
 import { populateDb } from "server/models/seed-data/populate-database";
 import rateLimit from "express-rate-limit";
 import {
